[ONNX] Update LayerNorm symbolic to handle autocasting (#52199)
When onnx export creates a 0-dim tensor of constant type, this action overrides the type promotion logic as quoted in #9515. In order to prevent this from happening this PR adds the following functionality.
If the data type is a floating point type, it is converted to a 0-dim double tensor, else it is converted to a 0-dim tensor of its original type

ghstack-source-id: aafc6e4fd5e111914512f519ae8084ba05a585f5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52350